refactor(ScheduleOption): simplify logic flow
